Remove use of deprecated functions from `Dates`
Both `strptime` and `strftime` are deprecated in PHP 8.1. This commit
removes them and replaces them with the modern equivalents
(`date_parse_from_format` and `date`).

In order to accomodate this change, both `Dates::parseStrptime()` and
`Dates::strftime()` now require formats to be specified in the modern
DateTime format specified here:
https://www.php.net/manual/en/datetimeimmutable.createfromformat.php.
This is therefore a breaking change, even though the former method is
undocumented (as it is nonetheless public).